requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
